backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
boto3==1.38.0
python-dotenv==1.1.0
Pillow==11.2.1
requests==2.32.3
